Add ability to disconnect/reconnect
There seem to be minor timing issues when reconnecting,
since the header and payload are both 16 bytes and
interpreting one as the other messes things up. To fix
this, we can combine the message types into one so ordering
isn't an issue. For now the problem seems to have been solved
by removing the 5ms delay between sending the header and payload.
I'm not sure why this works though.

Also removed the NWThread class. Instead a new thread is started
using the @run_async decorator every time we connect. Threads using
old socket information return automatically when the sockets are
closed.