Change how we read packets on UDP.
Read some stuff and learned that UDP communicates through
datagrams, which are like discrete units. This is in contrast
to TCP, which uses streams. sample_server.py now sends the
header and a single payload in one datagram rather than in
two separate datagrams. networker.py reads the header and all
payload data in a single datagram. The method of reading a
header then data doesn't work for UDP (I think, which explains
the errors I was getting previously).